chore(lints): deny `clippy::as_conversions` package-wide
A numeric `as` narrows silently: `65536u32 as u16` is 0 and nothing says so. The parser modules take their lengths from a peer, so the conversion that has to be checked is exactly the one an `as` lets somebody write without thinking. The lint is denied for every target in the package, beside `unsafe_code` and in the same shape: deny by default, and a site that wants one argues for its own allow with a reason a reader can check.

`src/` had 70 of them. 28 became conversions: `usize::from` and `u64::from` where the widening has an impl, `usize::try_from(..).unwrap_or(..)` with the saturation stated where it does not, `(0u16..).zip(..)` and `(0u64..).zip(..)` where the index was being counted in the wrong type to begin with, and `MAX_CODE_BITS` in the Huffman decoder is now the `usize` every one of its uses wanted. 42 stay, under 26 allows: three whole functions whose job is byte-level packing (`datagram::put_varint`, `qpack::put_int`, `net::blocking_pool_size`), the two `FrameDecoder` arms whose `take` is bounded by the chunk it came from, and the rest per statement. No bound in `src/h3` moved, nothing that was buffered stopped being buffered, and no refusal path was added: every remaining cast is one a line above it has already put in range.

`tests/` had 170, across 25 files, and each of those files carries a file-level allow with one line saying why: a test writes the numbers it asserts on. `fuzz/` is a separate workspace and inherits no `[lints]` from the root, the way its own `unsafe_code` stanza already records, so the lint does not reach it and it is left alone.

+++ src/auth.rs
@@ -409,14 +409,16 @@ mod tests {
 
         let mut out = String::new();
         for quantum in input.chunks(3) {
-            let mut bits: u32 = 0;
+            // Accumulated as a `usize` because every sextet is an index into
+            // `ALPHABET`; three bytes is 24 bits, well inside one.
+            let mut bits: usize = 0;
             for i in 0..3 {
-                bits = (bits << 8) | u32::from(quantum.get(i).copied().unwrap_or(0));
+                bits = (bits << 8) | usize::from(quantum.get(i).copied().unwrap_or(0));
             }
             let significant = quantum.len() + 1;
             for i in 0..significant {
                 let sextet = (bits >> (18 - 6 * i)) & 0x3f;
-                out.push(char::from(ALPHABET[sextet as usize]));
+                out.push(char::from(ALPHABET[sextet]));
             }
             for _ in significant..4 {
                 out.push('=');
